@@ -7,14 +7,12 @@
   },
   "license": "AGPL-3.0",
   "engines": {
-    "node": "8.11.3",
-    "npm": "5.6.0"
+    "node": "12.x"
   },
   "main": "js/app.js",
   "dependencies": {
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "^7.4.4",
-    "@babel/polyfill": "^7.4.4",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@sentry/browser": "^5.20.0",
@@ -122,78 +94,48 @@
     "react-select": "^3.0.8",
     "react-select-search": "^0.9.3",
     "react-selectize": "^2.1.0",
-    "react-simple-pie-chart": "^0.4.1",
     "react-sortable": "^2.0.0",
     "react-sortable-hoc": "^1.11.0",
     "react-table": "^6.7.5",
     "react-textarea-autosize": "^4.3.2",
-    "react-timezone": "^2.3.0",
     "react-tooltip": "^4.1.2",
-    "react-transition-group": "^1.2.1",
-    "react-wakelock": "0.0.4",
     "react-wakelock-react16": "0.0.7",
     "redux": "^3.3.1",
     "redux-persist": "^4.10.2",
     "redux-thunk": "^2.0.1",
     "request": "^2.81.0",
     "request-promise": "^4.2.2",
-    "resolve-url-loader": "^3.1.0",
-    "sass-loader": "^7.1.0",
+    "sass-loader": "^10.0.2",
     "sass-resources-loader": "^2.0.1",
     "sbd": "^1.0.18",
     "socket.io-client": "^2.3.0",
     "string-hash": "^1.1.3",
     "string-strip-html": "^4.3.22",
-    "style-loader": "^0.23.1",
-    "uglifyjs-webpack-plugin": "^2.2.0",
+    "style-loader": "^1.2.1",
     "underscore": "^1.8.3",
     "underscore.string": "^3.3.5",
-    "url-loader": "^2.0.1",
+    "url-loader": "^4.1.0",
     "uuid4": "^2.0.2",
     "warning": "^3.0.0",
-    "webpack": "^4.32.2",
-    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack": "^4.44.1",
     "webpack-manifest-plugin": "^2.0.4",
-    "webpack-merge": "^4.1.3",
+    "webpack-merge": "^5.1.3",
     "xlsx": "^0.14.5"
   },
   "devDependencies": {
     "@types/jest": "^22.2.0",
-    "acorn": "^6.4.1",
-    "acorn-dynamic-import": "^4.0.0",
-    "assets-webpack-plugin": "^3.9.12",
-    "awesome-typescript-loader": "^5.2.1",
     "babel-jest": "^24.8.0",
-    "babel-plugin-react-transform": "^2.0.2",
-    "body-parser": "^1.15.0",
-    "browserify-cipher": "^1.0.1",
-    "compression-webpack-plugin": "^3.1.0",
+    "compression-webpack-plugin": "^5.0.2",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "enzyme-to-json": "^3.4.4",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^14.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^3.0.1",
-    "eslint-plugin-react": "^6.9.0",
-    "estraverse-fb": "^1.3.1",
     "expect": "^1.20.2",
-    "expect-jsx": "^2.6.0",
-    "faye-websocket": "^0.11.1",
-    "hard-source-webpack-plugin": "^0.5.18",
+    "expose-loader": "^1.0.0",
+    "file-loader": "^6.1.0",
     "jest": "^24.8.0",
     "jest-localstorage-mock": "^2.4.3",
-    "jsdom": "^11.11.0",
-    "jsdom-global": "^3.0.2",
-    "json-loader": "^0.5.7",
-    "memory-fs": "^0.4.1",
-    "mime": "^2.2.0",
     "mocha": "^6.2.3",
     "qs": "^6.9.2",
-    "react-hot-loader": "^4.12.21",
-    "react-radio-group": "^3.0.3",
-    "react-test-renderer": "^16.13.0",
-    "react-transform-hmr": "^1.0.4",
     "redux-devtools": "^3.5.0",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-extension": "^2.13.8",
