@@ -7,8 +7,7 @@
   },
   "license": "AGPL-3.0",
   "engines": {
-    "node": "8.11.3",
-    "npm": "5.6.0"
+    "node": "12.x"
   },
   "cacheDirectories": [
     "client/node_modules"
@@ -48,30 +47,21 @@
     "rendering"
   ],
   "dependencies": {
-    "faker": "^4.1.0",
     "foreman": "^3.0.1",
-    "gitbook-plugin-rss": "git+https://git@github.com/ddmck/gitbook-plugin-rss.git",
     "isomorphic-fetch": "^2.2.1",
-    "istanbul-lib-instrument": "^1.10.2",
     "lazysizes": "^5.2.0-beta1",
     "quill-component-library": "0.10.2",
     "react-csv": "^1.1.2",
-    "react-dropdown": "^1.6.4",
     "react-radio-buttons": "^1.2.1",
     "react-select": "^1.1.0",
     "react-table": "^6.10.3",
-    "thenby": "^1.3.0",
-    "unfetch": "^4.1.0"
+    "thenby": "^1.3.0"
   },
   "devDependencies": {
     "@types/mocha": "^7.0.2",
-    "faye-websocket": "^0.11.3",
-    "memory-fs": "^0.4.1",
-    "mime": "^2.4.4",
     "mocha": "^7.1.1",
     "qs": "^6.9.0",
-    "redux-test-utils": "^0.2.2",
-    "ts-node": "^8.8.2"
+    "redux-test-utils": "^0.2.2"
   },
   "jest": {
     "roots": [
