@@ -5,7 +5,7 @@
 // cd client && yarn run build:client
 // Note that Foreman (Procfile.dev) has also been configured to take care of this.
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 const config = require('./webpack.client.dev.base.config');
 const { resolve } = require('path');
 const webpackConfigLoader = require('react-on-rails/webpackConfigLoader');
@@ -24,13 +24,6 @@ if (devBuild) {
 
 module.exports = merge(config, {
 
-  entry: {
-    vendor: [
-      // Configures extractStyles to be true if NODE_ENV is production
-      'bootstrap-loader'
-    ],
-  },
-
   mode: devBuild ? 'development' : 'production',
 
   output: {
@@ -41,7 +34,9 @@ module.exports = merge(config, {
   },
 
   optimization: {
-    splitChunks: false,
+    splitChunks: {
+      name: false
+    },
     removeAvailableModules: false,
     removeEmptyChunks: false,
     minimize: false,
@@ -66,25 +61,6 @@ module.exports = merge(config, {
           'css-loader',
           'sass-loader'
         ],
-      },
-      {
-        test: require.resolve('react'),
-        use: {
-          loader: 'imports-loader',
-          options: {
-            shim: 'es5-shim/es5-shim',
-            sham: 'es5-shim/es5-sham',
-          }
-        }
-      },
-      {
-        test: require.resolve('jquery-ujs'),
-        use: {
-          loader: 'imports-loader',
-          options: {
-            jQuery: 'jquery',
-          }
-        }
       }
     ],
   },
