@@ -5,15 +5,14 @@
 // cd client && yarn run build:client
 // Note that Foreman (Procfile.dev) has also been configured to take care of this.
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const merge = require('webpack-merge');
+const TerserPlugin = require('terser-webpack-plugin');
+const { merge } = require('webpack-merge');
 const config = require('./webpack.client.base.config');
 const { resolve } = require('path');
 const webpackConfigLoader = require('react-on-rails/webpackConfigLoader');
-const TerserPlugin = require('terser-webpack-plugin');
 
 const configPath = resolve('..', 'config');
 const { output } = webpackConfigLoader(configPath);
-
 const devBuild = process.env.RAILS_ENV === 'development';
 
 if (devBuild) {
@@ -25,13 +24,6 @@ if (devBuild) {
 
 module.exports = merge(config, {
 
-  entry: {
-    vendor: [
-      // Configures extractStyles to be true if NODE_ENV is production
-      'bootstrap-loader'
-    ],
-  },
-
   mode: devBuild ? 'development' : 'production',
 
   output: {
@@ -42,28 +34,13 @@ module.exports = merge(config, {
 
   optimization: {
     splitChunks: {
-      chunks: 'async',
-      minSize: 30000,
-      maxSize: 0,
-      minChunks: Infinity,
-      maxAsyncRequests: 5,
-      maxInitialRequests: 3,
-      automaticNameDelimiter: '~',
-      name: true,
-      cacheGroups: {
-        vendors: {
-          test: /[\\/]node_modules[\\/]/,
-          priority: -10
-        },
-        default: {
-          minChunks: Infinity,
-          priority: -20,
-          reuseExistingChunk: true
-        }
-      }
+      name: false
     },
     minimizer: [
-      new TerserPlugin(),
+      new TerserPlugin({
+        parallel: true,
+        cache: true
+      }),
     ],
   },
 
@@ -75,8 +52,7 @@ module.exports = merge(config, {
         test: /\.css$/,
         use: [
           MiniCssExtractPlugin.loader,
-          'css-loader',
-          'sass-loader'
+          'css-loader'
         ],
       },
       {
@@ -86,25 +62,6 @@ module.exports = merge(config, {
           'css-loader',
           'sass-loader'
         ],
-      },
-      {
-        test: require.resolve('react'),
-        use: {
-          loader: 'imports-loader',
-          options: {
-            shim: 'es5-shim/es5-shim',
-            sham: 'es5-shim/es5-sham',
-          }
-        }
-      },
-      {
-        test: require.resolve('jquery-ujs'),
-        use: {
-          loader: 'imports-loader',
-          options: {
-            jQuery: 'jquery',
-          }
-        }
       }
     ],
   },
@@ -115,6 +72,6 @@ module.exports = merge(config, {
       // both options are optional
       filename: '[name]-bundle-[hash].css',
       chunkFilename: '[id].css',
-    }),
+    })
   ],
-});
+})
