@@ -101,7 +101,7 @@ describe 'SerializeVitallySalesAccount' do
       classroom_unit: classroom_unit,
       state: 'finished'
     )
-    create(:activity_session,
+    last_activity_session = create(:activity_session,
       user: active_student,
       classroom_unit: classroom_unit,
       state: 'finished'
@@ -116,5 +116,6 @@ describe 'SerializeVitallySalesAccount' do
       active_students: 1,
       activities_finished: 2
     )
+    expect(school_data[:traits][:last_active]).to be_within(0.000001.second).of(last_activity_session.completed_at)
   end
 end
